feat(next): extract next-auth-redis to dedicated generator
