+++ packages/next/src/generators/next-auth-redis/files/src/index.test.ts__template__
