+++ packages/next/src/generators/next-auth-redis/files/src/index.ts__template__
