+++ packages/next/src/generators/next-auth-redis/schema.d.ts
@@ -1,0 +1,5 @@
+export interface NextAuthRedisGeneratorSchema {
+    project: string;
+    adapterName?: string;
+    envVar?: string;
+}
