+++ packages/next/src/generators/next-auth-redis/utils/constants.ts
@@ -1,0 +1,5 @@
+export const IOREDIS_VERSION = '^5.3.0';
+export const UUID_VERSION = '^9.0.0';
+export const TYPES_UUID_VERSION = '^9.0.0';
+
+export const DEFAULT_REDIS_URL = 'localhost:6379';
