+++ packages/next/src/generators/next-auth-redis/utils/update-helm-templates.ts
