+++ packages/next/src/generators/next-auth-redis/utils/update-terraform-files.ts
