+++ packages/next/src/generators/next-auth/schema.d.ts
@@ -1,8 +1,5 @@
 export interface NextAuthGeneratorSchema {
     project: string;
     provider: 'none' | 'azureAd' | 'azureAdB2C';
-    redisAdapter?: boolean;
-    redisAdapterName?: string;
-    redisEnvVar?: string;
     skipPackageJson?: boolean;
 }
