+++ packages/next/src/generators/next-auth/schema.json
@@ -23,22 +23,6 @@
             "default": "none",
             "x-prompt": "Which provider do you want to add?"
         },
-        "redisAdapter": {
-            "type": "boolean",
-            "default": false,
-            "description": "Add Redis adapter to store sessions?"
-        },
-        "redisEnvVar": {
-            "type": "string",
-            "default": "REDIS_URL",
-            "description": "Name of the env var that stores connection string for Redis",
-            "x-prompt": "What is the name of the env var that stores connection string for Redis?"
-        },
-        "redisAdapterName": {
-            "type": "string",
-            "default": "next-auth-redis",
-            "description": "Name of the generated Redis adapter package"
-        },
         "skipPackageJson": {
             "type": "boolean",
             "default": false,
