+++ packages/next/src/generators/next-auth/utils/constants.ts
@@ -1,6 +1,1 @@
 export const NEXT_AUTH_VERSION = '4.18.8';
-export const IOREDIS_VERSION = '^5.3.0';
-export const UUID_VERSION = '^9.0.0';
-export const TYPES_UUID_VERSION = '^9.0.0';
-
-export const DEFAULT_REDIS_URL = 'localhost:6379';
